+++ app/lib/pages/apps/app_detail.dart
@@ -7,7 +7,6 @@ import 'package:friend_private/backend/preferences.dart';
 import 'package:friend_private/pages/apps/analytics.dart';
 import 'package:friend_private/pages/apps/markdown_viewer.dart';
 import 'package:friend_private/pages/apps/providers/add_app_provider.dart';
-import 'package:friend_private/pages/apps/update_app.dart';
 import 'package:friend_private/pages/apps/widgets/info_card_widget.dart';
 import 'package:friend_private/providers/app_provider.dart';
 import 'package:friend_private/providers/connectivity_provider.dart';
@@ -20,138 +19,7 @@ import 'package:provider/provider.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../../backend/schema/app.dart';
-
-class ShowAppOptionsSheet extends StatelessWidget {
-  final App app;
-  const ShowAppOptionsSheet({super.key, required this.app});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Theme.of(context).colorScheme.surface,
-        borderRadius: const BorderRadius.only(topLeft: Radius.circular(16), topRight: Radius.circular(16)),
-      ),
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-      child: Consumer<AppProvider>(builder: (context, provider, child) {
-        return Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            ListTile(
-              title: Text(
-                app.name,
-                style: Theme.of(context).textTheme.labelLarge,
-              ),
-              leading: const Icon(Icons.apps),
-              trailing: IconButton(
-                icon: const Icon(Icons.cancel_outlined),
-                onPressed: () {
-                  Navigator.of(context).pop(true);
-                },
-              ),
-            ),
-            Card(
-              shape: const RoundedRectangleBorder(
-                borderRadius: BorderRadius.all(Radius.circular(8)),
-              ),
-              child: ListTile(
-                title: const Text(
-                  'Keep App Public',
-                  style: TextStyle(color: Colors.white, fontSize: 16, fontWeight: FontWeight.w600),
-                ),
-                trailing: Switch(
-                  value: provider.appPublicToggled,
-                  onChanged: (value) {
-                    if (value) {
-                      showDialog(
-                        context: context,
-                        builder: (c) => getDialog(
-                          context,
-                          () => Navigator.pop(context),
-                          () {
-                            provider.toggleAppPublic(app.id, value);
-                            Navigator.pop(context);
-                            Navigator.pop(context);
-                            Navigator.pop(context);
-                          },
-                          'Make App Public?',
-                          'If you make the app public, it can be used by everyone',
-                          okButtonText: 'Confirm',
-                        ),
-                      );
-                    } else {
-                      showDialog(
-                        context: context,
-                        builder: (c) => getDialog(
-                          context,
-                          () => Navigator.pop(context),
-                          () {
-                            provider.toggleAppPublic(app.id, value);
-                            Navigator.pop(context);
-                            Navigator.pop(context);
-                            Navigator.pop(context);
-                          },
-                          'Make App Private?',
-                          'If you make the app private now, it will stop working for everyone and will be visible only to you',
-                          okButtonText: 'Confirm',
-                        ),
-                      );
-                    }
-                  },
-                ),
-              ),
-            ),
-            Card(
-              shape: const RoundedRectangleBorder(
-                borderRadius: BorderRadius.all(Radius.circular(8)),
-              ),
-              child: Column(
-                children: [
-                  ListTile(
-                    title: const Text('Update App Details'),
-                    leading: const Icon(Icons.edit),
-                    onTap: () {
-                      routeToPage(
-                          context,
-                          UpdateAppPage(
-                            app: app,
-                          ));
-                    },
-                  ),
-                  ListTile(
-                    title: const Text('Delete App'),
-                    leading: const Icon(
-                      Icons.delete,
-                    ),
-                    onTap: () {
-                      showDialog(
-                        context: context,
-                        builder: (c) => getDialog(
-                          context,
-                          () => Navigator.pop(context),
-                          () {
-                            provider.deleteApp(app.id);
-                            Navigator.pop(context);
-                            Navigator.pop(context);
-                            Navigator.pop(context);
-                          },
-                          'Delete App',
-                          'Are you sure you want to delete this app? This action cannot be undone.',
-                          okButtonText: 'Confirm',
-                        ),
-                      );
-                    },
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 40),
-          ],
-        );
-      }),
-    );
-  }
-}
+import 'widgets/show_app_options_sheet.dart';
 
 class AppDetailPage extends StatefulWidget {
   final App app;
@@ -213,7 +81,6 @@ class _AppDetailPageState extends State<AppDetailPage> {
                 ? IconButton(
                     icon: const Icon(Icons.settings),
                     onPressed: () async {
-                      // open bottom sheet
                       await showModalBottomSheet(
                         context: context,
                         shape: const RoundedRectangleBorder(
@@ -383,7 +250,7 @@ class _AppDetailPageState extends State<AppDetailPage> {
                         child: AnimatedLoadingButton(
                           text: 'Uninstall App',
                           width: MediaQuery.of(context).size.width * 0.9,
-                          onPressed: () => _toggleApp(widget.app.id, false, context),
+                          onPressed: () => _toggleApp(widget.app.id, false),
                           color: Colors.red,
                         ),
                       ),
@@ -394,7 +261,7 @@ class _AppDetailPageState extends State<AppDetailPage> {
                         child: AnimatedLoadingButton(
                           width: MediaQuery.of(context).size.width * 0.9,
                           text: 'Install App',
-                          onPressed: () => _toggleApp(widget.app.id, true, context),
+                          onPressed: () => _toggleApp(widget.app.id, true),
                           color: Colors.green,
                         ),
                       ),
@@ -686,9 +553,38 @@ class _AppDetailPageState extends State<AppDetailPage> {
         ));
   }
 
-  Future<void> _toggleApp(String appId, bool isEnabled, BuildContext context) async {
+  Future<void> _toggleApp(String appId, bool isEnabled) async {
+    var prefs = SharedPreferencesUtil();
     setState(() => appLoading = true);
-    await context.read<AppProvider>().toggleApp(appId, isEnabled, null);
+    if (isEnabled) {
+      var enabled = await enableAppServer(appId);
+      if (!enabled) {
+        if (mounted) {
+          showDialog(
+            context: context,
+            builder: (c) => getDialog(
+              context,
+              () => Navigator.pop(context),
+              () => Navigator.pop(context),
+              'Error activating the app',
+              'If this is an integration app, make sure the setup is completed.',
+              singleButton: true,
+            ),
+          );
+        }
+
+        setState(() => appLoading = false);
+        return;
+      }
+
+      prefs.enableApp(appId);
+      MixpanelManager().appEnabled(appId);
+    } else {
+      prefs.disableApp(appId);
+      await enableAppServer(appId);
+      MixpanelManager().appDisabled(appId);
+    }
+    context.read<AppProvider>().setApps();
     setState(() => widget.app.enabled = isEnabled);
     setState(() => appLoading = false);
   }
